third_party/WebKit/LayoutTests/fast/dom/custom/type-extension-is-null.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
    var A = document.registerElement('x-foo', {extends: null, prototype: null});
    assert_equals(new A().outerHTML, '<x-foo></x-foo>',
        'localName should not be "null" and type extension should not have "is" attribute');

    assert_equals(new A().localName, 'x-foo',
        'localName should be set to TYPE if extends is null');

}, 'type extension from null should be treated as if "extends" is not given');
</script>
